element > EvenFirst

+++ elements/CountOne.java
@@ -4,10 +4,10 @@
 */
 public class CountOne {
 	static int count(int n) {
-		int count = 0;
+		short count = 0;
 		while(n != 0) {
 			count += (n & 1);
-			n >>>= 1;
+			n >>>= 1; 
 		}
 		return count;
 	}
